python/requirements.txt
# Core dependencies
python-dotenv==1.0.0

# PostHog
posthog>=6.6.1

# Provider-specific dependencies
# Anthropic
anthropic>=0.18.0

# Google Gemini
google-genai

# LangChain
langchain
langchain-core
langchain-openai

# OpenAI (also needed for LangChain)
openai

# LiteLLM
litellm

# LlamaIndex (for OpenTelemetry testing)
llama-index
llama-index-llms-openai

# OpenTelemetry (for LlamaIndex tracing to PostHog)
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http

# OpenInference (OpenTelemetry instrumentation for LlamaIndex)
openinference-instrumentation-llama-index
openinference-semantic-conventions

# OpenAI instrumentation for OpenTelemetry
opentelemetry-instrumentation-openai-v2